requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
python-dotenv>=1.0.0
requests>=2.31.0
